+++ romio/adio/common/shfp_fname.c
@@ -1,6 +1,6 @@
 /* -*- Mode: C; c-basic-offset:4 ; -*- */
 /* 
- *   $Id: shfp_fname.c,v 1.4 2002/10/24 17:01:15 gropp Exp $    
+ *   $Id: shfp_fname.c,v 1.5 2003/04/24 21:10:20 David Exp $    
  *
  *   Copyright (C) 1997 University of Chicago. 
  *   See COPYRIGHT notice in top-level directory.
@@ -58,7 +58,7 @@ void ADIOI_Shfp_fname(ADIO_File fd, int rank)
 	sprintf(tmp, ".shfp.%d", i);
 	strcat(fd->shared_fp_fname, tmp);
 	
-	len = strlen(fd->shared_fp_fname);
+	len = (int)strlen(fd->shared_fp_fname);
 	MPI_Bcast(&len, 1, MPI_INT, 0, fd->comm);
 	MPI_Bcast(fd->shared_fp_fname, len+1, MPI_CHAR, 0, fd->comm);
     }
